This pull request builds on #2 and adds a module for the [opentitan hmac HWIP](https://docs.opentitan.org/hw/ip/hmac/doc/).

## Current progress:

- [x] implement SHA256 part of hmac module
- [x] implement reading of cycle count  
- [x] fix outstanding bugs 

## Outstanding bugs:

[fixed] When compiling using rust's release mode, the resulting binary does not behave in the same way. Preliminary tests pass on debug builds (standard) but not on optimized builds.

## Preliminary results:

Using verilator to simulate the architecture, and running the suite compiled in release mode yields the following output:
```
Sha256 example-runtime analysis of 2048 bits:
initialization: 33 cycles
writing_input:  1125 cycles
computation*:   162 cycles
reading_output: 56 cycles
```
Note that the hmac HWIP performs partial computation while the input is written.  \
Also there are busy wait loops during writing of the input, that are used to wait for the hmac HWIP to consume the current data, so the rest of the input can be written. \
Hence the distinction between `writing_input` and `computation` is not accurate.